Reimplementing problem10 solver using sumValues

+++ src/katas/refactor/euler/Incrementor.java
@@ -4,14 +4,14 @@ import java.util.Vector;
 
 public abstract class Incrementor {
 
-	public abstract int increment(int current_value,
-			Vector<Integer> previous_values);
+	public abstract long increment(long current_value,
+			Vector<Long> previous_values);
 
 	public static Incrementor incrementBy(final int increment) {
 		return new Incrementor() {
 			@Override
-			public int increment(int current_value,
-					Vector<Integer> previous_values) {
+			public long increment(long current_value,
+					Vector<Long> previous_values) {
 				return current_value + increment;
 			}
 		};
@@ -20,8 +20,8 @@ public abstract class Incrementor {
 	public static Incrementor sumLastTwoValuesOr(final int defaultValue) {
 		return new Incrementor() {
 			@Override
-			public int increment(int current_value,
-					Vector<Integer> previous_values) {
+			public long increment(long current_value,
+					Vector<Long> previous_values) {
 				int size = previous_values.size();
 				return (size == 1) ? defaultValue : previous_values
 						.get(size - 1) + previous_values.get(size - 2);
