+++ src/katas/refactor/euler/ProblemSolver1.java
@@ -16,7 +16,7 @@ public class ProblemSolver1 extends ProblemSolver {
 		return "Add all the natural numbers below one thousand that are multiples of 3 or 5.";
 	}
 
-	protected Integer[] getFirstXNaturals(final int x) {
+	protected Long[] getFirstXNaturals(final int x) {
 		return getFirstNumbers(0, Incrementor.incrementBy(1),
 				Condition.lessThan(x));
 	}
